sign in and sign out test updated
